Add widgetState only stories
