+++ assets/javascripts/kitten/components/cards/crowdfunding-card/components/state.js
@@ -2,9 +2,8 @@ import React, { PureComponent } from 'react'
 import styled from 'styled-components'
 import PropTypes from 'prop-types'
 import COLORS from '../../../../constants/colors-config'
-import { Text } from '../../../../components/typography/text'
 import { pxToRem } from '../../../../helpers/utils/typography'
-import { KissKissBankBankIcon } from '../../../../components/icons/kisskissbankbank-icon'
+import { Text } from '../../../../components/typography/text'
 
 const COMPONENT_GUTTER = pxToRem(10)
 
@@ -28,15 +27,6 @@ const StyledWidgetContainer = styled.div`
   right: 0;
   top: 0;
   padding: ${pxToRem(10)} ${pxToRem(20)};
-`
-
-const StyledWidgetText = styled(Text)`
-  display: flex;
-  align-items: center;
-`
-
-const StyledWidgetLogo = styled(KissKissBankBankIcon)`
-  margin-left: ${pxToRem(5)};
 `
 
 class State extends PureComponent {
@@ -90,17 +80,7 @@ class State extends PureComponent {
     return (
       <>
         {!loading && (
-          <StyledWidgetContainer>
-            <StyledWidgetText
-              size="micro"
-              lineHeight="normal"
-              weight="regular"
-              color="background1"
-            >
-              {widgetState}
-              <StyledWidgetLogo width="80" height="18" />
-            </StyledWidgetText>
-          </StyledWidgetContainer>
+          <StyledWidgetContainer>{widgetState}</StyledWidgetContainer>
         )}
       </>
     )
